+++ env_base.php
@@ -4,3 +4,8 @@
     define('DISCORD_TOKEN','');
     define('DISCORD_CHANNEL','');
     define('DISCORD_CHANNEL_ADMIN','');
+    define('USE_OPENAI',0);
+    define('OPENAI_API_KEY','');
+    define('OPENAI_ASSISTANT_CONTEXT','');
+    define('OPENAI_ASSISTANT_ID','');
+    define('OPENAI_MODEL','');